+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ### Changelog:
 - ServerChatEvent for messages sent to the player on the Server!
+- Made damage event cancel any incoming damage if return value is 0.0F